Modifs pour MSVC

+++ Base/include/private/DBaseImageOperations.hxx
@@ -53,7 +53,7 @@ inline typename Image<T>::lineType *imageFunctionBase<T>::createAlignedBuffers(U
     bufferSize = bufferLength * sizeof(T);
 
     alignedBuffers = new lineType[bufferNumber];
-    for (int i=0;i<bufferNumber;i++)
+    for (size_t i=0;i<bufferNumber;i++)
         alignedBuffers[i] = ImDtTypes<T>::createLine(len);
 
     return alignedBuffers;
@@ -148,11 +148,11 @@ inline RES_T binaryImageFunction<T, lineFunction_T>::_exec(const imageType &imIn
     lineType *srcLines2 = imIn2.getLines();
     lineType *destLines = imOut.getLines();
 
-    UINT i, chunk = 100;
+    int i, chunk = 100;
     #pragma omp parallel shared(srcLines1,srcLines2,destLines,chunk) private(i)
     {
 	#pragma omp for schedule(dynamic,chunk) nowait
-	for (i=0;i<lineCount;i++)
+	for (i=0;i<(int)lineCount;i++)
 	    lineFunction(srcLines1[i], srcLines2[i], lineLen, destLines[i]);
     }
     imOut.modified();
